Debug ministage 2
