+++ Werewolf/WP/__init__.py
@@ -5,4 +5,4 @@ The protocol for the werewolf game
 import os
 import socket
 import sys
-from .api import ChunckedData, ReceiveThread, _recv, TimeLock
+from .api import ChunckedData, ReceiveThread, _recv, TimeLock, KillableThread, ReadInput
