+++ Werewolf/server/logic.py
@@ -841,45 +841,3 @@ class IncomingConnection(Thread):
             print(c.getsockname())
             print("Client connected")
             self.game.addPlayer(c, _recv(c))
-
-
-class KillableThread(Thread):
-    """
-    A thread class extending threading.Thread, provides a kill() method to stop the thread and a getResult() method to get the return value of the thread.
-    """
-
-    def __init__(self, func: Callable, **kwargs):
-        super().__init__()
-        self.func: Callable = func
-        self.funcArg: dict = kwargs
-        self.result: Any = None
-
-    def run(self):
-        """
-        Executes the function here
-        """
-        self.result = self.func(**self.funcArg)
-
-    def get_id(self):
-        """
-        Get the id of the thread
-        """
-        if hasattr(self, '_thread_id'):
-            return self._thread_id
-        for id, thread in threading._active.items():
-            if thread is self:
-                return id
-
-    def kill(self):
-        """
-        Stops the thread
-        """
-        thread_id = self.get_id()
-        res = ctypes.pythonapi.PyThreadState_SetAsyncExc(thread_id,
-                                                         ctypes.py_object(SystemExit))
-        if res > 1:
-            ctypes.pythonapi.PyThreadState_SetAsyncExc(thread_id, 0)
-            print('Exception raise failure')
-
-    def getResult(self):
-        return self.result
